Refactored Code - Removed redundant functions - refactored variables into more meaningful names - Reordered import statements
