+++ src/app/app.component.html
@@ -3,9 +3,9 @@
       <button mat-icon-button (click)="sidenav.toggle()" fxShow="true" fxHide.gt-sm>
         <mat-icon>menu</mat-icon>
       </button>
-      <span (click)='goHome()' [routerLink]=''>Countdown Timer</span>
+      <span [routerLink]="'/'">{{title}}</span>
       <span class="flex-spacer"></span>
-      <div *ngIf="auth.user$ | async as user; else login" fxShow="true" fxHide.lt-md="true">
+      <div *ngIf="auth.user$ | async as user; else navBarNoAuth" fxShow="true" fxHide.lt-md="true">
         <a mat-button [routerLink]="'/account'"> Hello, {{ user.displayName }} </a>
         <a mat-button (click)="auth.signOut()"> Sign Out </a>
         <a mat-button [routerLink]="'/events'"> Events </a>
@@ -19,7 +19,7 @@
 
 <mat-sidenav-container fxFlexFill>
   <mat-sidenav #sidenav>
-    <mat-nav-list *ngIf="auth.user$ | async as user; else loginNav">
+    <mat-nav-list *ngIf="auth.user$ | async as user; else sideNavNoAuth">
       <a (click)="sidenav.toggle()" href="" mat-list-item>Close</a>
       <a mat-button [routerLink]="'/'"> Home </a>
       <a mat-button [routerLink]="'/account'"> Hello, {{ user.displayName }} </a>
@@ -36,7 +36,7 @@
   </mat-sidenav-content>
 </mat-sidenav-container>
 
-<ng-template #login>
+<ng-template #navBarNoAuth>
   <div fxShow="true" fxHide.lt-md="true">
     <a mat-button (click)="auth.googleSignin()"> Log In </a>
     <a mat-button [routerLink]="'/events'"> Events </a>
@@ -44,7 +44,7 @@
   </div>
 </ng-template>
 
-<ng-template #loginNav>
+<ng-template #sideNavNoAuth>
   <mat-nav-list>
     <a (click)="sidenav.toggle()" href="" mat-list-item>Close</a>
     <a mat-button [routerLink]="'/'"> Home </a>
