+++ src/app/components/carousel/carousel.component.html
@@ -2,7 +2,7 @@
   <div class="carousel-inner">
     <ul id="init-card" class="carousel-item active">
       <!-- <img src="..." class="d-block w-100" alt="..."> -->
-      <app-featured-card [data]='init'></app-featured-card>
+      <app-featured-card [data]='previewCard'></app-featured-card>
     </ul>
     <ul  *ngFor='let item of data' class="carousel-item" id="{{ item.id }}">
       <app-featured-card [data]='item'></app-featured-card>
