+++ src/app/components/carousel/carousel.component.ts
@@ -1,7 +1,8 @@
 import { Component, OnInit } from '@angular/core';
 
 import { FirebaseServiceService } from '../../services/firebase-service.service';
-import { Event } from '../../models/event';
+
+import { Event } from '../../models/event.model';
 
 @Component({
   selector: 'app-carousel',
@@ -12,7 +13,7 @@ import { Event } from '../../models/event';
   ]
 })
 export class CarouselComponent implements OnInit {
-  init: Event = {};
+  previewCard: Event = {};
   data: Event[] = [];
 
   constructor(private firebaseService: FirebaseServiceService) { }
@@ -22,7 +23,7 @@ export class CarouselComponent implements OnInit {
       this.data = items;
     });
     
-    this.init = {
+    this.previewCard = {
       title: 'The Next Day',
       subtitle: 'Preview Card',
       content: 'This is a Preview card, with Dummy data. This is kept here to check if all things are loaded and displayed properly. This card will be removed when Development is complte and is Deployed.',
