+++ src/app/components/featured-card/featured-card.component.ts
@@ -1,7 +1,9 @@
 import { Component, OnInit, Input } from '@angular/core';
+
 import { FirebaseServiceService } from "../../services/firebase-service.service";
 
-import { Event } from '../../models/event';
+import { Event } from '../../models/event.model';
+
 import { incrementCount, copyID } from '../../utils'
 
 @Component({
